move and refactor Traces components to playground-ui/domains
