+++ packages/playground-ui/src/domains/traces/components/format-hierarchical-spans.ts
