+++ packages/playground-ui/src/domains/traces/components/timeline-expand-col.tsx
@@ -1,5 +1,5 @@
-import { cn } from '@mastra/playground-ui';
 import { ChevronDownIcon, ChevronsDownIcon, ChevronsUpIcon, ChevronUpIcon } from 'lucide-react';
+import { cn } from '@/lib/utils';
 
 type TimelineExpandColProps = {
   isSelected?: boolean;
