+++ packages/playground-ui/src/domains/traces/components/timeline-name-col.tsx
@@ -1,6 +1,6 @@
-import { cn } from '@mastra/playground-ui';
 import type { UISpan, UISpanStyle } from '../types';
 import { TimelineStructureSign } from './timeline-structure-sign';
+import { cn } from '@/lib/utils';
 
 type TimelineNameColProps = {
   span: UISpan;
