+++ packages/playground-ui/src/domains/traces/hooks/use-span-detail.ts
