+++ packages/playground-ui/src/domains/traces/hooks/use-trace-light-spans.ts
