+++ packages/playground-ui/src/domains/traces/hooks/use-traces.tsx
@@ -1,8 +1,9 @@
 import type { ListTracesArgs, ListTracesResponse } from '@mastra/core/storage';
-import { useInView, is403ForbiddenError } from '@mastra/playground-ui';
 import { useMastraClient } from '@mastra/react';
 import { useInfiniteQuery, keepPreviousData } from '@tanstack/react-query';
 import { useEffect } from 'react';
+import { useInView } from '@/hooks/use-in-view';
+import { is403ForbiddenError } from '@/lib/query-utils';
 
 const fetchTracesFn = async ({
   client,
