+++ packages/playground-ui/src/domains/traces/index.ts
@@ -1,0 +1,9 @@
+export * from './components';
+export * from './hooks';
+export * from './utils';
+export * from './trace-filters';
+export type { UISpan, UISpanStyle, TraceDatePreset, EntityOptions } from './types';
+export { CONTEXT_FIELD_IDS } from './types';
+
+/** Tab identifier for SpanDataPanelView. */
+export type SpanTab = 'details' | 'scoring';
