+++ packages/playground-ui/src/domains/traces/utils/group-traces-by-thread.ts
