+++ packages/playground-ui/src/domains/traces/utils/span-utils.ts
