+++ packages/playground/src/domains/logs/components/trace/timeline-expand-col.tsx
@@ -1,1 +1,1 @@
-export { TimelineExpandCol } from '@/domains/traces/components/timeline-expand-col';
+export { TimelineExpandCol } from '@mastra/playground-ui';
