+++ packages/playground/src/domains/logs/components/trace/timeline-timing-col.tsx
@@ -1,1 +1,1 @@
-export { TimelineTimingCol } from '@/domains/traces/components/timeline-timing-col';
+export { TimelineTimingCol } from '@mastra/playground-ui';
