+++ packages/playground/src/domains/logs/components/trace/trace-timeline-span.tsx
@@ -1,1 +1,1 @@
-export { TraceTimelineSpan } from '@/domains/traces/components/trace-timeline-span';
+export { TraceTimelineSpan } from '@mastra/playground-ui';
